NetBox v4 Update
