+++ netbox_topology_plugin/constants/icons.py
@@ -1,0 +1,21 @@
+SUPPORTED_ICONS = {
+    'switch',
+    'router',
+    'firewall',
+    'wlc',
+    'unknown',
+    'server',
+    'phone',
+    'nexus5000',
+    'ipphone',
+    'host',
+    'camera',
+    'accesspoint',
+    'groups',
+    'groupm',
+    'groupl',
+    'cloud',
+    'unlinked',
+    'hostgroup',
+    'wirelesshost',
+}
