+++ netbox_topology_plugin/constants/interfaces.py
@@ -1,0 +1,6 @@
+INTERFACE_FULL_NAME_MAP = {
+    'Eth': 'Ethernet',
+    'Fa': 'FastEthernet',
+    'Gi': 'GigabitEthernet',
+    'Te': 'TenGigabitEthernet',
+}
