+++ netbox_topology_plugin/templates/netbox_topology_plugin/site_topo_button.html
@@ -1,4 +1,4 @@
-<button type="button" class="btn btn-primary" data-toggle="modal" data-target="#topology_modal" data-obj="{{ object.name }}" title="Show site topology">
+<button type="button" class="btn btn-primary" data-bs-toggle="modal" data-bs-target="#topology_modal" data-bs-obj="{{ object.name }}" title="Show site topology">
     <i class="fa fa-signal" aria-hidden="true"></i>
     Topology
 </button>
@@ -7,8 +7,8 @@
     <div class="modal-dialog modal-lg" role="document">
         <div class="modal-content">
                 <div class="modal-header">
-                    <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
-                    <h2 class="modal-title" style="text-align: center;"><a href={% url 'plugins:netbox_topology_plugin:topology' %}?site_id={{ object.id }}>{{ object.name }} Topology</a></h2>
+                    <button type="button" class="btn btn-sm btn-danger" data-bs-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
+                    <h2 class="modal-title" style="text-align: center;width:100%;"><a href={% url 'plugins:netbox_topology_plugin:topology' %}?site_id={{ object.id }}>{{ object.name }} Topology</a></h2>
                 </div>
         <iframe src="{% url 'plugins:netbox_topology_plugin:site_topology' %}?site_id={{ object.id }}" style="height: 75vh; width: 100%;"></iframe>
         </div>
