+++ netbox_topology_plugin/templates/netbox_topology_plugin/site_topology.html
@@ -1,61 +1,51 @@
+{% extends 'base/base.html' %}
 {% load static %}
 
-<!DOCTYPE html>
-
-<html>
-    <head>
-        <meta charset="utf-8">
+{% block head %}
         <link rel="stylesheet"href="{% static 'netbox_topology_plugin/next_sources/css/next.css' %}">
-        <link rel="stylesheet"href="{% static 'netbox_topology_plugin/bootstrap.min.css' %}">
-        <link rel="stylesheet"href="{% static 'font-awesome-4.7.0/css/font-awesome.min.css' %}">
         <script src="{% static 'netbox_topology_plugin/next_sources/js/next.js' %}"></script>
         <script src="{% static 'netbox_topology_plugin/button_utils.js' %}"></script>
-    </head>
-    <body>
+
+{% endblock %} 
+
+{% block layout %}
+
         <div class="container-sm ml-4 my-sm-3" id="nx-ui-topology">
-            <div class="row">
-              <div class="col-xs-4">
-                <div class="btn-group">
-                    <button class="btn btn-outline-secondary btn-sm " onclick='horizontal()'>Horizontal Layout</button>
-                    <button class="btn btn-outline-secondary btn-sm" onclick="vertical()">Vertical Layout</button>
-                </div>
+            <div class="control-group">
+              <div class="btn-group">
+                  <button class="btn btn-primary btn-sm " onclick='horizontal()'>Horizontal Layout</button>
+                  <button class="btn btn-primary btn-sm" onclick="vertical()">Vertical Layout</button>
               </div>
-              <div class="col-xs-4 ml-2">
-                <input class="btn btn-outline-secondary btn-sm" type="button" id="showHideUndonnectedButton" value="" onclick="return showHideUndonnectedButtonOnClick(this);" />
-              </div>
-              <div class="col-xs-4 ml-2">
-                <input class="btn btn-outline-secondary btn-sm" type="button" id="showHidePassiveDevicesButton" value="" onclick="return showHidePassiveDevicesButtonOnClick(this);" />
-              </div>
-              <div class="col-xs-4 ml-2">
-                <div class="dropdown">
-                    <button class="btn btn-outline-secondary btn-sm dropdown-toggle" type="button" id="dropdownMenuButton" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-                      Select Layers
-                    </button>
-                    <div class="dropdown-menu" aria-labelledby="dropdownMenuButton">
-                        <h6 class="dropdown-header">Device Roles</h6>
-                        <div class="dropdown-divider"></div>
-                        {% for device_role_slug, device_role_name, is_visible in device_roles %}
+              <input class="btn btn-primary btn-sm" type="button" id="showHideUndonnectedButton" value="" onclick="return showHideUndonnectedButtonOnClick(this);" />
+              <input class="btn btn-primary btn-sm" type="button" id="showHidePassiveDevicesButton" value="" onclick="return showHidePassiveDevicesButtonOnClick(this);" />
+              <div class="dropdown btn-group">
+                  <button class="btn btn-primary btn-sm dropdown-toggle" type="button" id="dropdownMenuButton" data-bs-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+                    Select Layers
+                  </button>
+                  <div class="dropdown-menu" aria-labelledby="dropdownMenuButton">
+                      <h6 class="dropdown-header">Device Roles</h6>
+                      <div class="dropdown-divider"></div>
+                      {% for device_role_slug, device_role_name, is_visible in device_roles %}
+                          <div class="checkbox ml-1 my-1">
+                              <label>
+                                  <input onchange="layerSelectorOnChange(this)" type="checkbox" value="{{ device_role_slug }}" {{ is_visible|yesno:"checked," }}>{{ device_role_name }}
+                              </label>
+                          </div>
+                      {% endfor %}
+                      <div class="dropdown-divider"></div>
+                      {% if device_tags %}
+                        <h6 class="dropdown-header">Device Tags</h6>
+                        {% for tag, is_visible in device_tags %}
                             <div class="checkbox ml-1 my-1">
                                 <label>
-                                    <input onchange="layerSelectorOnChange(this)" type="checkbox" value="{{ device_role_slug }}" {{ is_visible|yesno:"checked," }}>{{ device_role_name }}
+                                   <input onchange="layerSelectorByTagOnChange(this)" type="checkbox" value="{{ tag }}"  {{ is_visible|yesno:"checked," }}>{{ tag }}
                                 </label>
                             </div>
                         {% endfor %}
-                        <div class="dropdown-divider"></div>
-                        {% if device_tags %}
-                          <h6 class="dropdown-header">Device Tags</h6>
-                          {% for tag, is_visible in device_tags %}
-                              <div class="checkbox ml-1 my-1">
-                                  <label>
-                                      <input onchange="layerSelectorByTagOnChange(this)" type="checkbox" value="{{ tag }}" {{ is_visible|yesno:"checked," }}>{{ tag }}
-                                  </label>
-                              </div>
-                          {% endfor %}
-                        {% endif %}
-                        <div class="dropdown-divider"></div>
-                    </div>
+                      {% endif %}
+                      <div class="dropdown-divider"></div>
                   </div>
-              </div>
+                </div>
             </div>
         </div>
         
@@ -74,9 +64,5 @@
             showHidePassiveDevicesButtonInitial();
         </script>
         <script src="{% static 'netbox_topology_plugin/next_app.js' %}"></script>
-        <script src="{% static 'jquery/jquery-3.4.1.min.js' %}"></script>
-        <script src="{% static 'jquery/jquery-3.5.1.min.js' %}"></script>
-        <script src="{% static 'jquery-ui-1.12.1/jquery-ui.min.js' %}"></script>
-        <script src="{% static 'netbox_topology_plugin/bootstrap.bundle.min.js' %}"></script>
-    </body>
-</html>
+
+{% endblock %}
