+++ netbox_topology_plugin/utils/filtering.py
@@ -1,0 +1,42 @@
+import re
+
+from netbox_topology_plugin.settings import UNDISPLAYED_DEVICE_TAGS, SELECT_LAYERS_LIST_INCLUDE_DEVICE_TAGS, \
+    SELECT_LAYERS_LIST_EXCLUDE_DEVICE_TAGS
+
+
+__all__ = (
+    'filter_tags',
+    'tag_is_hidden',
+)
+
+
+def tag_is_hidden(tag):
+    for tag_regex in UNDISPLAYED_DEVICE_TAGS:
+        if re.search(tag_regex, tag):
+            return True
+    return False
+
+
+def filter_tags(tags):
+    if not tags:
+        return []
+    if SELECT_LAYERS_LIST_INCLUDE_DEVICE_TAGS:
+        filtered_tags = []
+        for tag in tags:
+            for tag_regex in SELECT_LAYERS_LIST_INCLUDE_DEVICE_TAGS:
+                if re.search(tag_regex, tag):
+                    filtered_tags.append(tag)
+                    break
+            if tag_is_hidden(tag):
+                filtered_tags.append(tag)
+        tags = filtered_tags
+    if SELECT_LAYERS_LIST_EXCLUDE_DEVICE_TAGS:
+        filtered_tags = []
+        for tag in tags:
+            for tag_regex in SELECT_LAYERS_LIST_EXCLUDE_DEVICE_TAGS:
+                if re.search(tag_regex, tag) and not tag_is_hidden(tag):
+                    break
+            else:
+                filtered_tags.append(tag)
+        tags = filtered_tags
+    return tags
