+++ pyproject.toml
@@ -15,7 +15,7 @@ description = "Netbox Topology Plugin, fork of netbox-topology-plugin.  Powered 
 readme = "README.md"
 requires-python = ">=3.9"
 keywords = ["netbox-plugin", "netbox-topology"]
-version = "0.14.2"
+version = "0.14.3"
 license = {file = "LICENSE"}
 classifiers = [
     "Programming Language :: Python :: 3",
